% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_doublets.R
\name{find_doublets}
\alias{find_doublets}
\title{Find and Visualize Doublets in Subclusters of a Seurat Object}
\usage{
find_doublets(
  so = seurat_object,
  cell_type_col = "cell_types",
  ndim = 20,
  var_regress = "percent.mt",
  harmony_vars = c("sample"),
  clust_res = 0.1,
  doublet_scores_col = "scDbt_score",
  output = "subcluster_doublets"
)
}
\arguments{
\item{so}{A Seurat object containing single-cell RNA-seq data.}

\item{cell_type_col}{A string specifying the column name in the Seurat object metadata
that contains cell type labels. Default is "cell_types".}

\item{ndim}{An integer specifying the number of principal components to compute for
clustering and UMAP. Default is 20.}

\item{var_regress}{A string specifying the variable to regress out using SCTransform.
Default is "percent.mt".}

\item{harmony_vars}{A character vector specifying the variables to integrate using harmony.
Default is c("sample").}

\item{clust_res}{A numeric value specifying the resolution for subclustering. Default is 0.1.}

\item{doublet_scores_col}{A string specifying the column name in the Seurat object metadata
that contains doublet scores from scDblFinder. Default is "scDbt_score".}

\item{output}{A string specifying the directory to save output plots. Default is "subcluster_doublets".}
}
\value{
A character vector of subcluster identifiers.
}
\description{
This function performs sub-clustering of cell types within a Seurat object,
identifies potential doublets using doublet scores, and generates visualizations
for subclusters and doublet scores.
}
